TOR management is based on OVS protocol. A new role called TSN (TOR services node)
is introduced for compute node. TSN handles multicast / broadcast traffic from TOR,
along with DHCP and DNS services. A new process contrail-tor-agent will be run for
every TOR that is managed. This process handles configuring TOR using OVSDB, as also
exporting the routes reported by the TOR.

Gateway Port on compute node can be configured with a subnet that is reachable on
the port.

+++ src/base/test/task_test_util.h
@@ -90,7 +90,7 @@ do {                                                                           \
 #define TASK_UTIL_EXPECT_VECTOR_EQ(actual, expected)             \
     do {                                                         \
         TASK_UTIL_EXPECT_EQ((expected).size(), (actual).size()); \
-        for (int i = 0; i < (expected).size(); i++) {            \
+        for (int i = 0; i < (int)((expected).size()); i++) {            \
             TASK_UTIL_EXPECT_EQ((expected)[i], (actual)[i]);     \
         }                                                        \
     } while (false)
